Guard against REMOVE messages for data: URLs for tunnelled dialogs in iOS app
Trying to remove such a "file" caused a crash.

Sure, loleaflet should ideally not send such REMOVE messages at all.

No idea why the image for a tunnelled dialog shows up as a data: URL
and not as a file: URL pointing to a BMP file, like the document
tiles.

+++ ios/Mobile/DocumentViewController.mm
@@ -527,6 +527,12 @@ static IMP standardImpOfInputAccessoryView = nil;
             NSArray<NSString*> *messageBodyItems = [message.body componentsSeparatedByString:@" "];
             assert([messageBodyItems count] == 2);
             NSURL *tile = [NSURL URLWithString:messageBodyItems[1]];
+
+            // For some reason tunnelled dialogs still use PNG tiles inside data: URLs and not BMP
+            // files pointed to by file: URLs. Guard against getting REMOVE messages for such.
+            if (![[tile scheme] isEqualToString:@"file"])
+                return;
+
             if (unlink([[tile path] UTF8String]) == -1) {
                 LOG_SYS("Could not unlink tile " << [[tile path] UTF8String]);
             }
